modified searchbar again, added subroutes for recipe categories, added pathname component
