+++ src/components/ui/button.tsx
@@ -1,5 +1,5 @@
 import { type ComponentProps } from "react";
-import cn from "./cn";
+import cn from "../cn";
 
 type ButtonProps = {
   variant?: "normal" | "ghost";
